indexing: update mef persons mapping
* Adds analyzer on biographical_information, preferred_name_for_person,
  authorized_access_point_representing_a_person and variant_name_for_person fields.
